+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [3.5.0] - 2023-03-15
+- Replaced the use of arg_parser with gen_class.ArgParser class
+- Upgraded python-lib to v2.10.1
+
+### Changed
+- Multiple functions: Replaced the use of arg_parser (args_array) with gen_class.ArgParser class (args).
+
+
 ## [3.4.4] - 2022-12-15
 - Updated to work in Python 3 too
 - Upgraded python-lib to v2.9.4